ref(control_silo): Consolidate integrations web resources
Moves all integrations web resources and related tests to sentry/integrations/web and tests/sentry/integrations/web

ref(#73859)
