+++ src/sentry/integrations/web/debug/debug_notify_disable.py
@@ -4,8 +4,7 @@ from django.views.generic import View
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.notify_disable import get_provider_type, get_url
 from sentry.models.organization import Organization
-
-from .mail import MailPreview
+from sentry.web.frontend.debug.mail import MailPreview
 
 
 class DebugNotifyDisableView(View):
