+++ src/sentry/integrations/web/discord_extension_configuration.py
