+++ src/sentry/integrations/web/doc_integration_avatar.py
