+++ src/sentry/integrations/web/integration_extension_configuration.py
