+++ src/sentry/integrations/web/vercel_extension_configuration.py
