+++ tests/sentry/integrations/web/test_msteams_extension.py
@@ -2,12 +2,14 @@ from unittest.mock import patch
 
 from django.core.signing import SignatureExpired
 
+from sentry.integrations.web.msteams_extension_configuration import (
+    MsTeamsExtensionConfigurationView,
+)
 from sentry.models.organizationmember import OrganizationMember
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
 from sentry.utils.signing import sign
-from sentry.web.frontend.msteams_extension_configuration import MsTeamsExtensionConfigurationView
 
 
 @control_silo_test
